Town Map update (+Makefile and Readme)
Visual changes to the Town Map, as well as an expansion for the tileset. If needed, it can be expanded further, but it will need to be loaded at a prior address (currently loaded at $30).

Readme was updated to point the RGBDS version used, as well as remove mentions to pokered and linking to pret's discord. Can be updated later with more info and screenshots.

Makefile was updated to only build two roms (comet and comet_debug). They're .sgb files, so they were added to .gitignore. They're based on Blue's but have an unique header.