Tileset Expansion
This commit expands the tileset, allowing one full row and one partial row (up to $77) to be used.

A new gfx for the tree stump was added as well.

Some maps were already update, namely the starting town and route 1, other maps need to be updated.

Added a docs folder to put any docs about the project.